traffic_generator/requirements.txt
requests
psycopg2-binary
matplotlib